pov: reduce the flywheel column with a 32-bit modulo (finding 38)
floor_mod(int64_t, int32_t) promoted the modulus to 64 bits, so
Flywheel::position() emitted two __aeabi_ldivmod calls on the 18.4 kHz
flywheel wake. Narrow the column count to int32 (the signed-safe window
assert already bounds it to MIN_SAFE_HALF_REVS half-revs of columns) and
add an int32 floor_mod overload; the second library call becomes a
hardware sdiv. The 64-bit overload keeps its semantics for other callers.

+++ hardware/pov_sync_flywheel.h
@@ -101,7 +101,10 @@ public:
          elapsed >= 0u - static_cast<uint32_t>(MIN_SAFE_HALF_REVS) * period) &&
         "Flywheel::position: timestamp outside the signed-safe coast window");
     const int64_t delta = static_cast<int32_t>(elapsed);
-    const int64_t cols = floor_div(delta * (w / 2), period);
+    // |cols| <= MIN_SAFE_HALF_REVS*(W/2) under the window assert, so the
+    // narrowing is exact and the reduction stays a 32-bit sdiv.
+    const int32_t cols =
+        static_cast<int32_t>(floor_div(delta * (w / 2), period));
     return floor_mod(boundary_column(boundary, w) + cols, w);
   }
